zbxjira/meus_requirements.txt
jira==1.0.10
zabbix-api
pyyaml
sys
os
argparse
